Fixes for PR: renamed each component/page within its own folder to index.js; re-structured import chain of scss files.
