+++ src/components/common/SocialMediaContainer/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import 'components/common/SocialMediaContainer/social-media-container.scss';
 import { ReactComponent as Iphone } from 'assets/i6.svg';
-import appStoreImg from 'assets/appstore.png'; // with import
-import socialLinksImg from 'assets/socialnetworks.png';
+import appStoreImg from 'assets/appstore.png';
+import fbIconImg from 'assets/facebook-icon.png';
+import twitterIconImg from 'assets/twitter-icon.png';
 
 const SocialMediaContainer = () => {
   return (
@@ -11,13 +11,16 @@ const SocialMediaContainer = () => {
         <Iphone />
       </div>
       <div className="app-store-container">
-        <a href="">
+        <a href="https://www.apple.com/">
           <img src={appStoreImg} alt="App Store" />
         </a>
       </div>
       <div className="social-links-container">
-        <a href="">
-          <img src={socialLinksImg} alt="Social Links" />
+        <a href="https://www.facebook.com/">
+          <img src={fbIconImg} alt="Facebook Icon" />
+        </a>
+        <a href="https://www.twitter.com/">
+          <img src={twitterIconImg} alt="Twitter Icon" />
         </a>
       </div>
     </div>
