+++ src/components/user/LoginForm/index.js
@@ -10,8 +10,6 @@ import { login as loginValidations } from 'utils/constraints';
 import Loading from 'components/common/Loading';
 import Input from 'components/common/Input';
 
-import './login-form.scss';
-
 const messages = defineMessages({
   email: { id: 'login.form.email' },
   password: { id: 'login.form.password' }
