+++ src/pages/LoginPage/login-page.scss
@@ -1,5 +1,3 @@
-@import 'src/styles/_variables.scss';
-
 .login-page-container {
   display: flex;
   min-height: 100vh;
